filenames: extension logic fix
Sadly, `.set_extension` replaces a file name extension, it does
not append, and newer interfaces to append extensions are not
yet stable.  Anyway, this just modifies some logic so that we
append the `.efs.json5` extension when constructing a config
filename when not explicitly passing the path to the config.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>